chore: Move OAH* non-performance-critical methods to the .cc file.
